Version 0.0.12
- Modified SearchEngine (pending removing searchResults and adding a
SearchPOJO)
- Added SearchPOJO
- Moved Element<T> to a separate package
