+++ ingsoft/src/main/java/es/ucm/fdi/business/util/Element.java
@@ -1,0 +1,29 @@
+package es.ucm.fdi.business.util;
+
+/**
+ * This class is designed to hold search results, indicating whether a possible match is
+ * or not visible in the current search.
+ *
+ * @param <T> Type of the content of the class
+ */
+public class Element<T>{
+	private boolean visible;
+	private T element;
+	
+	public Element(T e){
+		visible = false;
+		element = e;
+	}
+	
+	public boolean isVisible(){
+		return visible;
+	}
+	
+	public T getElement(){
+		return element;
+	}
+	
+	public void setVisible(boolean b){
+		visible = b;
+	}
+}
